feat: seperation of manager into seperate managers
Purpose: to seperate the control of each manager type from one another
