+++ pkg/controllers/managerClusterManaged.go
@@ -1,0 +1,156 @@
+package controllers
+
+import (
+	"context"
+
+	bootstrapController "github.com/ksctl/ksctl/pkg/controllers/bootstrap"
+	cloudController "github.com/ksctl/ksctl/pkg/controllers/cloud"
+	"github.com/ksctl/ksctl/pkg/helpers"
+	"github.com/ksctl/ksctl/pkg/helpers/consts"
+	"github.com/ksctl/ksctl/pkg/types"
+	storageTypes "github.com/ksctl/ksctl/pkg/types/storage"
+)
+
+type ManagerClusterManaged struct {
+	managerInfo
+}
+
+func NewManagerClusterManaged(ctx context.Context, log types.LoggerFactory, client *types.KsctlClient) (*ManagerClusterManaged, error) {
+	defer panicCatcher(log)
+
+	stateDocument = new(storageTypes.StorageDocument)
+	controllerCtx = context.WithValue(ctx, consts.KsctlModuleNameKey, "ksctl-manager")
+
+	cloudController.InitLogger(controllerCtx, log)
+	bootstrapController.InitLogger(controllerCtx, log)
+
+	manager := new(ManagerClusterManaged)
+	manager.log = log
+	manager.client = client
+
+	err := manager.initStorage(controllerCtx)
+	if err != nil {
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return nil, err
+	}
+
+	return manager, nil
+}
+
+func (manager *ManagerClusterManaged) CreateCluster() error {
+	client := manager.client
+	log := manager.log
+	defer panicCatcher(log)
+
+	if err := manager.setupConfigurations(); err != nil {
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	if client.Metadata.Provider == consts.CloudLocal {
+		client.Metadata.Region = "LOCAL"
+	}
+
+	if err := client.Storage.Setup(
+		client.Metadata.Provider,
+		client.Metadata.Region,
+		client.Metadata.ClusterName,
+		consts.ClusterTypeMang); err != nil {
+
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	defer func() {
+		if err := client.Storage.Kill(); err != nil {
+			log.Error(controllerCtx, "StorageClass Kill failed", "reason", err)
+		}
+	}()
+
+	fakeClient := false
+	if _, ok := helpers.IsContextPresent(controllerCtx, consts.KsctlTestFlagKey); ok {
+		fakeClient = true
+	}
+
+	if !helpers.ValidCNIPlugin(consts.KsctlValidCNIPlugin(client.Metadata.CNIPlugin)) {
+		err := log.NewError(controllerCtx, "invalid CNI plugin")
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	if err := cloudController.InitCloud(client, stateDocument, consts.OperationCreate, fakeClient); err != nil {
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	// it gets supportForApps, supportForCNI, error
+	externalApp, externalCNI, cloudResErr := cloudController.CreateManagedCluster(client)
+	if cloudResErr != nil {
+		log.Error(controllerCtx, "handled error", "catch", cloudResErr)
+		return cloudResErr
+	}
+
+	if err := bootstrapController.InstallAdditionalTools(
+		externalCNI,
+		externalApp,
+		client,
+		stateDocument); err != nil {
+
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	log.Success(controllerCtx, "successfully created managed cluster")
+	return nil
+}
+
+func (manager *ManagerClusterManaged) DeleteCluster() error {
+
+	client := manager.client
+	log := manager.log
+
+	defer panicCatcher(log)
+	if err := manager.setupConfigurations(); err != nil {
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	if client.Metadata.Provider == consts.CloudLocal {
+		client.Metadata.Region = "LOCAL"
+	}
+
+	if err := client.Storage.Setup(
+		client.Metadata.Provider,
+		client.Metadata.Region,
+		client.Metadata.ClusterName,
+		consts.ClusterTypeMang); err != nil {
+
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	defer func() {
+		if err := client.Storage.Kill(); err != nil {
+			log.Error(controllerCtx, "StorageClass Kill failed", "reason", err)
+		}
+	}()
+
+	fakeClient := false
+	if _, ok := helpers.IsContextPresent(controllerCtx, consts.KsctlTestFlagKey); ok {
+		fakeClient = true
+	}
+
+	if err := cloudController.InitCloud(client, stateDocument, consts.OperationDelete, fakeClient); err != nil {
+		log.Error(controllerCtx, "handled error", "catch", err)
+		return err
+	}
+
+	cloudResErr := cloudController.DeleteManagedCluster(client)
+	if cloudResErr != nil {
+		log.Error(controllerCtx, "handled error", "catch", cloudResErr)
+		return cloudResErr
+	}
+
+	log.Success(controllerCtx, "successfully deleted managed cluster")
+	return nil
+}
